+++ packages/@expo/cli/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 💡 Others
 
+- Drop outdated React Native resolver patch. ([#29214](https://github.com/expo/expo/pull/29214) by [@EvanBacon](https://github.com/EvanBacon))
 - Use Metro instance directly for server rendering. ([#28552](https://github.com/expo/expo/pull/28552) by [@EvanBacon](https://github.com/EvanBacon))
 
 ## 0.18.16 — 2024-06-06