+++ packages/@expo/cli/CHANGELOG.md
@@ -14,6 +14,7 @@
 ### 🐛 Bug fixes
 
 - Fix source map generation in development. ([#29463](https://github.com/expo/expo/pull/29463) by [@EvanBacon](https://github.com/EvanBacon))
+- Fix resolving the main entry file Expo Router static export on Windows
 
 ### 💡 Others